requirements.txt
bilibili-api-python>=17.0.0
aiohttp>=3.9.0
feedparser>=6.0.0